+++ examples/composition/note_pitch.cpp
@@ -5,11 +5,11 @@
 
 namespace barelyapi {
 
-float GetPitch(const std::vector<float>& scale, int index) noexcept {
+double GetPitch(const std::vector<double>& scale, int index) noexcept {
   assert(!scale.empty());
-  const float scale_length = static_cast<float>(scale.size());
-  const float octave_offset =
-      std::floor(static_cast<float>(index) / scale_length);
+  const double scale_length = static_cast<double>(scale.size());
+  const double octave_offset =
+      std::floor(static_cast<double>(index) / scale_length);
   const int scale_offset =
       index - static_cast<int>(octave_offset * scale_length);
